fix(sev): rename `enarx platform snp info`
Rename `enarx platform snp info` to `enarx platform snp vcek` to reflect
its real purpose. `enarx platform snp vcek --file` also sounds better.

Also extend the description of `enarx platform snp update`.
