+++ src/cli/platform/snp/mod.rs
@@ -1,21 +1,21 @@
 // SPDX-License-Identifier: Apache-2.0
 
-mod info;
 mod update;
+mod vcek;
 
 use clap::Subcommand;
 
 /// SNP-specific functionality
 #[derive(Subcommand, Debug)]
 pub enum Subcommands {
-    Info(info::Options),
+    Vcek(vcek::Options),
     Update(update::Options),
 }
 
 impl Subcommands {
     pub fn dispatch(self) -> anyhow::Result<()> {
         match self {
-            Self::Info(cmd) => cmd.execute(),
+            Self::Vcek(cmd) => cmd.execute(),
             Self::Update(cmd) => cmd.execute(),
         }
     }
