+++ src/cli/platform/snp/update.rs
@@ -4,9 +4,8 @@ use crate::backend::sev::snp::vcek::vcek_write;
 
 use clap::Args;
 
-/// Download the VCEK certificate for this platform.
-///
-/// The certificate will be saved to a cache file in `/var/cache/amd-sev/`
+/// Download the current VCEK certificate for this platform
+/// to a cache file in the `/var/cache/amd-sev/` directory
 #[derive(Args, Debug)]
 pub struct Options {}
