+++ src/cli/platform/snp/vcek.rs
